Extract FirefoxAccount's BrowserID methods in their own file
